backend readme edits (feel free to change)
Feel free to delete and edit as you wish! This is by no means a final project. I boiler plated some of the content from my previous build week. So some of the information is obviously wrong. But the front end read me is very up to date with our product plan. So feel free to reference that document to add notes to the back end read me if desired.